pyproject.toml
[tool.poetry]
name = "data-sampling-tools"
version = "0.3.0"
description = ""
authors = ["Andrei Margert <andrei.margert@wur.nl>"]
readme = "README.md"
packages = [{include = "echo_dataset"}]

[tool.poetry.dependencies]
python = ">=3.9,<3.11"
PyYAML = "^6.0"
schema = "^0.7.5"
xarray = "^2022.9.0"
zarr = "^2.13.2"
dask = "^2022.9.1"
opencv-python = "^4.6.0.66"
datatable = {git = "git@github.com:h2oai/datatable.git", branch = "main", python = ">3.9"}
holoviews = "^1.15.4"


[tool.poetry.group.dev.dependencies]
matplotlib = "^3.6.1"
black = {extras = ["d"], version = "^22.12.0"}
tqdm = "^4.64.1"
jupyter = "^1.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
